pyproject.toml
[project]
name = "aurora-mysql-encryption"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nikki Satmaka", email = "30800020+NikkiSatmaka@users.noreply.github.com" } 
]
requires-python = ">=3.12"
dependencies = [
    "cryptography>=43.0.1",
    "polars>=1.9.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
]

[project.scripts]
aurora-mysql-encryption = "aurora_mysql_encryption:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
